+++ pkg/streamingpromql/operators/selectors/selector.go
@@ -163,7 +163,7 @@ func (l *seriesList) Len() int {
 //
 // Calling ToSeriesMetadata after calling Pop may return an incomplete list.
 func (l *seriesList) ToSeriesMetadata() []types.SeriesMetadata {
-	metadata := types.GetSeriesMetadataSlice(l.length)
+	metadata := types.GetSeriesMetadataSlice(int64(l.length))
 	batch := l.currentSeriesBatch
 
 	for batch != nil {
